Remove language redirect from Affiliates links.
Because SUPPORTED_NONLOCALES doesn't allow paths with slashes in them,
the prefix on the banner app had to be removed to allow for the link url
to have a unique prefix.

The longterm solution is to move Affiliate link logic into an app itself, which
should be considered for a later fix.

bug 693376