Check for banner image before creating instance.
When an affiliate link is visited, the app will query for the specified
banner image before creating the bannerinstance. This fixes a bug wherein
a bannerinstance failed to be created due to a bad bannerimage id, but the
parent badgeinstance was created, causing an exception on the My Badges
page.

fix bug 693824