Allow BannerImages in any locale + English labels.
Changes LocaleField to allow any locale defined in product_details
to be used, making it possible to add BannerImages from locales
that the site has yet to be translated in.

In addition, this uses a custom ModelForm on every ModelAdmin object
across the site to change the labels for LocaleFields in the admin
interface to use English labels instead of localized labels.

fix bug 713041