Fixes an issue with unsupported locales.
Adds a product_language_lower dict that allows one to retrieve
the English and native names of a language using the lower-case
version of a locale code (as our database does) in order to
fix an issue in the customize view where it couldn't find
the name for locales that the site isn't fully localized
in.

Also changes the import name for the unfortunately-named
locale module.

fix bug 772882