Update banner list page with post to wall link.
Adds a link that redirects the user to the "Feed Dialog" so
they can post a banner to their Facebook Feed. After posting,
the user is redirected to the site, which in turn redirects
them back to the Facebook app (The feed dialog can only redirect
back to a URL the app owns, hence the double redirect).