Add admin interface for FacebookBannerInstances.
I'm not testing that FilterSpec. You may assume this is because
I am tired, and you would be partially correct. But the other
side of the story is that FilterSpecs are trickery intended to
get you to put field lookups in GET parameters. So the best we
could do is parse the HTML and make sure that the admin interface
has the right GET parameters in those links, and that when those 
links are visited, the filter links are colored correctly to show
which one was selected.

I've got better things to unit test. :D