Bug 771715: Add AppNotifications to tell users about special events.
Adds a new model, AppNotification, that is shown to the user once when
they login, and then is deleted. They're similar to django messages, but
are persistent.

Adds a hook in the click task to create notifications about click count
milestones, as well as a pre_save signal to FacebookBannerInstances that
tells users when their banner has been approved as an ad.

Also adds the ability to use fb_redirect for redirects within the app.
This solves an issue where context processors and middleware that
require an active login session wouldn't run on the first page because
the user auths during that request after they have run. Now, the first
page uses the JavaScript redirect in fb_redirect to force a new request
with the new authed session.