Fix bug 896423: Update our usage of the Facebook Send Dialog.
We actually are compliant with the October 2013 Breaking Changes, but
this removes the excess parameters anyway. This also fixes the 
post-invite view to correctly detect if the user sent a message or if
they cancelled out of the invitation process.

Also adds SITE_ID = 1 to the test settings to make the tests work even
if you've set SITE_ID to something else in your local settings. 
Otherwise the tests might attempt to fetch a SITE_ID besides 1, which
doesn't exist in the test database.