Update migration docs and add command for migrating data.
Instead of checking out certain commits and running migrations, instead
we're going to just move to the latest, make a few manual DB changes,
and pull in the old data from a copy of the old database.

No unit tests for the migration are included since the command will 
really only be used twice or so, and manual testing is far easier for
this particular case.