Bug 1007847: Migrate creation date for links.
Pulls the creation date for links during the migration process. Changes
the created field to use default=timezone.new instead of auto_now_add
so that the created date can be set when a link is created and not
overridden.

Also fixes an incorrectly frozen spec for the links table in order to
make schemamigration behave properly.