Bug 1008447: Add CachingMixin and CacheManager to User class.
When a new link is created, the cached results for a user’s links
doesn’t get invalidated because, while the link has a foreign key to the
user, the user isn’t cached by django-cache-machine. By adding the mixin
and manager to the user class, when a new link is created, the user, and
thus it’s related links, should be invalidated, fixing the caching 
issues on the dashboard.